requirements.txt
aiohttp==3.10.10
discord.py==2.4.0
DrissionPage==4.1.0.9
